Fix u64→u32 return types in 180 d5 batch functions
The original binary uses MOV W0, WZR (32-bit) to zero the return
register, but u64 return type generates MOV X0, XZR (64-bit).
Change return types from u64 to u32 to match.

+180 verified functions (6869 → 7049)

+++ src/app/fun_batch_d5_004.cpp
@@ -114,21 +114,21 @@ void FUN_7101523b40(void)
 // Pattern: call helper with *(param_1 - 8) + offset as argument
 
 // 0x7101f78db0 — FUN_71002f43c0(*(param_1 - 8) + 0x180)
-u64 FUN_7101f78db0(s64 param_1)
+u32 FUN_7101f78db0(s64 param_1)
 {
     FUN_71002f43c0(*(s64 *)(param_1 - 8) + 0x180);
     return 0;
 }
 
 // 0x7101f78dd0 — FUN_71002f4590(*(param_1 - 8) + 0x180)
-u64 FUN_7101f78dd0(s64 param_1)
+u32 FUN_7101f78dd0(s64 param_1)
 {
     FUN_71002f4590(*(s64 *)(param_1 - 8) + 0x180);
     return 0;
 }
 
 // 0x7101f78df0 — FUN_71002f4830(*(param_1 - 8) + 0x180)
-u64 FUN_7101f78df0(s64 param_1)
+u32 FUN_7101f78df0(s64 param_1)
 {
     FUN_71002f4830(*(s64 *)(param_1 - 8) + 0x180);
     return 0;
